LPS-65236 Move PortalSessionActivationListener to portal-kernel.
